requirements.txt
# apk2image
androguard==3.3.5
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==1.23.5
customtkinter==5.1.3
matplotlib==3.7.1
tensorflow==2.10.1
# jsons
os
tkinter
threading

# federated learning
flwr
flwr-datasets
torch
torchvision
matplotlib
ray
dotenv
